Kotlin 1.5 and dependencies updates

+++ mirai-core/src/commonMain/kotlin/utils/io/serialization/tars/internal/TarsDecoder.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019-2020 Mamoe Technologies and contributors.
+ * Copyright 2019-2021 Mamoe Technologies and contributors.
  *
  *  此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 许可证的约束, 可以在以下链接找到该许可证.
  *  Use of this source code is governed by the GNU AGPLv3 license that can be found through the following link.
@@ -20,7 +20,6 @@ import kotlinx.serialization.internal.TaggedDecoder
 import kotlinx.serialization.modules.SerializersModule
 import net.mamoe.mirai.internal.utils.io.serialization.tars.Tars
 import net.mamoe.mirai.internal.utils.io.serialization.tars.TarsId
-import net.mamoe.mirai.internal.utils.io.serialization.tars.internal.*
 
 
 @OptIn(InternalSerializationApi::class, ExperimentalSerializationApi::class)
@@ -110,7 +109,7 @@ internal class TarsDecoder(
         override fun decodeFloat(): Float = input.useHead { input.readTarsFloatValue(it) }
         override fun decodeDouble(): Double = input.useHead { input.readTarsDoubleValue(it) }
         override fun decodeBoolean(): Boolean = input.useHead { input.readTarsBooleanValue(it) }
-        override fun decodeChar(): Char = decodeByte().toChar()
+        override fun decodeChar(): Char = decodeByte().toInt().toChar()
         override fun decodeEnum(enumDescriptor: SerialDescriptor): Int = decodeInt()
         override fun decodeString(): String = input.useHead { input.readTarsStringValue(it) }
 
@@ -158,7 +157,7 @@ internal class TarsDecoder(
         override fun decodeDouble(): Double = input.useHead { input.readTarsDoubleValue(it) }
 
         override fun decodeBoolean(): Boolean = input.useHead { input.readTarsBooleanValue(it) }
-        override fun decodeChar(): Char = decodeByte().toChar()
+        override fun decodeChar(): Char = decodeByte().toInt().toChar()
         override fun decodeEnum(enumDescriptor: SerialDescriptor): Int = decodeInt()
         override fun decodeString(): String = input.useHead { head ->
             input.readTarsStringValue(head).also {
